requirements.txt
flask
python-binance
requests
pandas
python-telegram-bot